fix: `deserializeAsync` string chunks can split anywhere (#30)
String chunks can split the string anywhere and shouldn't be assumed to
imply a new line.

This PR insures that `deserializeAsync` always accumulates the *original
string* from the received chunks before splitting it into separate
lines. Before this change, it was assumed that "end of chunk" means "end
of line" even if the `\n` char wasn't in the original string.

It also adds a test in `deserializeAsync` that verifies it still works
with weird chunk splits. Test didn't pass before this PR, passes now.

---

Sorry for the branch name / PR formatting, I haven't contributed to tRPC
in a while I don't remember the conventions

---------

Co-authored-by: Alex / KATT <alexander@n1s.se>